fixed compile warnings

+++ myapp.h
@@ -32,7 +32,7 @@ public:
 	inline void CopyToSurface(Surface* screen, int offsetX, int offsetY) const {
 		for (auto y = 0; y < height; y++) {
 			for (auto x = 0; x < width; x++) {
-				screen->Plot(x + offsetX, y + offsetY, rgb2uint(pixels[y * width + x] / samples));
+				screen->Plot(x + offsetX, y + offsetY, rgb2uint(pixels[y * width + x] / (float)samples));
 			}
 		}
 	}
